--- packages/core/src/persisted/index.ts
@@ -1,2 +1,0 @@
-export { GraphQLPersisted } from './types';
-export { PersistedAutomatic } from './automatic';
