+++ packages/persisted/package.json
@@ -1,8 +1,8 @@
 {
-  "name": "@benzene/core",
-  "version": "0.1.1",
-  "description": "Fast and non-coupled JavaScript GraphQL Server and Executor",
-  "author": "Hoang Vo (https://hoangvvo.com)",
+  "name": "@benzene/persisted",
+  "version": "0.0.0",
+  "description": "Persisted queries for benzene",
+  "main": "dist/index.js",
   "repository": {
     "type": "git",
     "url": "git+https://github.com/hoangvvo/benzene.git"
@@ -11,7 +11,6 @@
     "url": "https://github.com/hoangvvo/benzene/issues"
   },
   "homepage": "https://hoangvvo.github.io/benzene/",
-  "main": "dist/index.js",
   "types": "dist/index.d.ts",
   "files": [
     "dist"
@@ -22,8 +21,8 @@
   },
   "license": "MIT",
   "dependencies": {
-    "@hoangvvo/graphql-jit": "^0.6.1",
-    "flatstr": "^1.0.12",
+    "@benzene/core": "^0.1.1",
+    "crypto-hash": "^1.3.0",
     "tiny-lru": "^7.0.6"
   },
   "peerDependencies": {
