+++ packages/persisted/src/index.ts
@@ -1,0 +1,1 @@
+export { default as PersistedAutomatic } from './automatic';
